backend/requirements.txt
anthropic==0.51.0
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0